a little refactory and homecrypt now use interctive
