+++ .oclif.manifest.json
@@ -1549,5 +1549,5 @@
       ]
     }
   },
-  "version": "25.10.27"
+  "version": "25.10.28"
 }
