+++ src/classes/ovary.d/luks-home.ts
@@ -78,14 +78,16 @@ export async function luksHome(
     warning(`partition LUKS ${this.luksFile} size: ${bytesToGB(luksSize)}`)
 
     warning(`creating partition LUKS: ${this.luksFile}`)
-    await executeCommand('truncate', ['--size', `${luksSize}`, this.luksFile])
+    await this.luksExecuteCommand('truncate', ['--size', `${luksSize}`, this.luksFile])
 
     warning(`formatting ${this.luksFile} as a LUKS volume...`)
-    await executeCommand('cryptsetup', ['--batch-mode', 'luksFormat', this.luksFile], `${this.luksPassword}\n`);
+    //await this.luksExecuteCommand('cryptsetup', ['--batch-mode', 'luksFormat', this.luksFile], `${this.luksPassword}\n`);
+    const luksFormatArgs = this.buildLuksFormatArgs(this.luksConfig, this.luksFile);
+    await this.luksExecuteCommand('cryptsetup', luksFormatArgs, `${this.luksPassword}\n`);
 
 
     warning(`opening the LUKS volume. It will be mapped to ${this.luksDevice}`)
-    await executeCommand('cryptsetup', ['luksOpen', this.luksFile, this.luksMappedName], `${this.luksPassword}\n`)
+    await this.luksExecuteCommand('cryptsetup', ['luksOpen', this.luksFile, this.luksMappedName], `${this.luksPassword}\n`)
 
     warning(`formatting c ext4 `)
     await exec(`mkfs.ext4 -L live-home ${this.luksDevice}`,this.echo)
@@ -133,7 +135,7 @@ export async function luksHome(
     await exec(`umount ${this.luksMountpoint}`, this.echo)
 
     warning(`closing LUKS volume ${this.luksMappedName}.`)
-    await executeCommand('cryptsetup', ['close', this.luksMappedName])
+    await this.luksExecuteCommand('cryptsetup', ['close', this.luksMappedName])
 
     warning(`moving ${this.luksMappedName}  to (ISO)/live/.`)
     await exec(`mv ${this.luksFile} ${this.settings.iso_work}/live`, this.echo)
@@ -152,43 +154,11 @@ export async function luksHome(
       await exec(`umount -lf ${this.luksMountpoint}`).catch(() => { })
     }
     if (fs.existsSync(this.luksDevice)) {
-      await executeCommand('cryptsetup', ['close', this.luksMappedName]).catch(() => { })
+      await this.luksExecuteCommand('cryptsetup', ['close', this.luksMappedName]).catch(() => { })
     }
     await Utils.pressKeyToExit()
     process.exit(1)
   }
-}
-
-/**
- * Funzione helper per eseguire comandi esterni in modo asincrono,
- * gestendo lo standard input per passare le password.
- * Restituisce una Promise che si risolve al successo o si rigetta in caso di errore.
- */
-function executeCommand(command: string, args: string[], stdinData?: string): Promise<void> {
-  return new Promise((resolve, reject) => {
-    // Se passiamo dati a stdin, dobbiamo usare 'pipe'. Altrimenti, 'inherit'.
-    const stdioConfig: StdioOptions = stdinData ? ['pipe', 'inherit', 'inherit'] : 'inherit';
-
-    const process = spawn(command, args, { stdio: stdioConfig });
-
-    // Se fornito, scriviamo i dati (es. la password) nello stdin del processo.
-    if (stdinData && process.stdin) {
-      process.stdin.write(stdinData);
-      process.stdin.end();
-    }
-
-    process.on('error', (err) => {
-      reject(new Error(`Error starting command "${command}": ${err.message}`));
-    });
-
-    process.on('close', (code) => {
-      if (code === 0) {
-        resolve(); // Success
-      } else {
-        reject(new Error(`Command "${command} ${args.join(' ')}" ended with error code ${code}`));
-      }
-    });
-  });
 }
