+++ src/classes/ovary.d/produce.ts
@@ -96,13 +96,8 @@ export async function produce(
         this.luksDevice = `/dev/mapper/${this.luksMappedName}`
         this.luksPassword = '0' // USARE UNA PASSWORD SICURA IN PRODUZIONE!
 
-        /**
-         * Al momento homecrypt resta con default
-         */
-        if (fullcrypt) {
-            this.luksConfig = await this.interactiveCryptoConfig()
-        }
         
+        this.luksConfig = await this.interactiveCryptoConfig()
 
         Utils.warning("You choose an encrypted eggs")
         await this.luksGetPassword()
